requirements.txt
# Core
python-dotenv>=1.0
pyyaml>=6.0

# LLM
openai>=1.0
anthropic>=0.18
langchain>=0.1
langchain-openai>=0.0.5
langchain-anthropic>=0.1

# NLP
spacy>=3.7

# PDF
pdfplumber>=0.10
PyPDF2>=3.0

# Dashboard
streamlit>=1.30

# Tokenization
tiktoken>=0.5

# Date parsing
python-dateutil>=2.8

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
